feat: make delete in drawer work refactor: factor out delete modal
