+++ frontend/src/components/pages/CampsList/index.tsx
@@ -1,9 +1,11 @@
-import { Flex, Box, useDisclosure } from "@chakra-ui/react";
+import { Flex, Box, useDisclosure, useToast } from "@chakra-ui/react";
 import React, { useState } from "react";
 import { CampResponse } from "../../../types/CampsTypes";
 import CampsNavigationHeading from "./CampsNavigationHeading";
 import CampsTable from "./CampsTable";
 import PreviewCampDrawer from "./PreviewCampDrawer";
+import CampsAPIClient from "../../../APIClients/CampsAPIClient";
+import DeleteModal from "../../common/DeleteModal";
 
 const CampsListPage = (): React.ReactElement => {
   const [year, setYear] = useState(new Date().getFullYear());
@@ -12,46 +14,113 @@ const CampsListPage = (): React.ReactElement => {
     onOpen: onDrawerOpen,
     onClose: onDrawerClose,
   } = useDisclosure();
+  const [camps, setCamps] = React.useState([] as CampResponse[]);
+
   const [campDrawerInfo, setCampDrawerInfo] = useState<CampResponse>();
+  const [campToDelete, setCampToDelete] = React.useState<CampResponse | null>(
+    null,
+  );
+  const { isOpen, onOpen, onClose } = useDisclosure();
+
+  const toast = useToast();
+
+  React.useEffect(() => {
+    const getCamps = async () => {
+      const res = await CampsAPIClient.getAllCamps(year);
+      if (res) {
+        setCamps(res);
+      }
+    };
+
+    getCamps();
+  }, [year]);
+
+  const handleConfirmDelete = async () => {
+    if (!campToDelete) {
+      return;
+    }
+
+    const res = await CampsAPIClient.deleteCamp(campToDelete.id);
+    if (res) {
+      toast({
+        description: `${campToDelete.name} has been succesfully deleted`,
+        status: "success",
+        variant: "subtle",
+        duration: 3000,
+      });
+
+      const newCampsList: CampResponse[] = await camps.filter((camp) => {
+        return camp.id !== campToDelete.id;
+      });
+      setCamps(newCampsList);
+    } else {
+      toast({
+        description: `An error occurred with deleting ${campToDelete.name}. Please try again.`,
+        status: "error",
+        variant: "subtle",
+        duration: 3000,
+      });
+    }
+    onClose();
+    setCampToDelete(null);
+  };
+
+  const onDeleteClick = (camp: CampResponse) => {
+    setCampToDelete(camp);
+    onOpen();
+  };
 
   return (
-    <Flex
-      width={isDrawerOpen ? "calc(100% - 500px)" : "100%"}
-      transition="width 0.5s"
-      top="68px"
-    >
-      <Box
-        minHeight="100vh"
-        px="3em"
-        py="3em"
-        background="background.grey.200"
-        flex="1"
-      >
-        <CampsNavigationHeading
-          year={year}
-          onNavigateLeft={() => {
-            setYear(year - 1);
-            onDrawerClose();
-          }}
-          onNavigateRight={() => {
-            setYear(year + 1);
-            onDrawerClose();
-          }}
-        />
-        <CampsTable
-          year={year}
-          isDrawerOpen={isDrawerOpen}
-          onDrawerOpen={onDrawerOpen}
-          campDrawerInfo={campDrawerInfo}
-          setCampDrawerInfo={setCampDrawerInfo}
-        />
-      </Box>
-      <PreviewCampDrawer
-        isOpen={isDrawerOpen}
-        onClose={onDrawerClose}
-        camp={campDrawerInfo}
+    <>
+      <DeleteModal
+        title="Delete Camp"
+        bodyText={`Are you sure you want to delete ${campToDelete?.name}?`}
+        bodyNote="Note: This action is irreversible."
+        buttonLabel="Remove"
+        isOpen={isOpen}
+        onClose={onClose}
+        onDelete={() => handleConfirmDelete()}
       />
-    </Flex>
+      <Flex
+        width={isDrawerOpen ? "calc(100% - 500px)" : "100%"}
+        transition="width 0.5s"
+        top="68px"
+      >
+        <Box
+          minHeight="100vh"
+          px="3em"
+          py="3em"
+          background="background.grey.200"
+          flex="1"
+        >
+          <CampsNavigationHeading
+            year={year}
+            onNavigateLeft={() => {
+              setYear(year - 1);
+              onDrawerClose();
+            }}
+            onNavigateRight={() => {
+              setYear(year + 1);
+              onDrawerClose();
+            }}
+          />
+          <CampsTable
+            camps={camps}
+            isDrawerOpen={isDrawerOpen}
+            onDrawerOpen={onDrawerOpen}
+            campDrawerInfo={campDrawerInfo}
+            setCampDrawerInfo={setCampDrawerInfo}
+            onDeleteClick={onDeleteClick}
+          />
+        </Box>
+        <PreviewCampDrawer
+          isOpen={isDrawerOpen}
+          onClose={onDrawerClose}
+          camp={campDrawerInfo}
+          onDeleteClick={onDeleteClick}
+        />
+      </Flex>
+    </>
   );
 };
